AW-112 rename rename white paper section to content highlight section (#97)
* feat (AW-116): rename to content offer

old white paper, todo remove old collection

* feat (AW-112): rename to content highlight section

old white paper, todo remove old section

* feat (AW-112): rename nextjs components and update middleware

* fix: bad merge

* feat AW-112: remove white paper

* fix: news schema

* feat AW-112: renamed
