+++ strapi/src/api/content-offer/controllers/content-offer.ts
@@ -1,0 +1,7 @@
+/**
+ * content-offer controller
+ */
+
+import { factories } from '@strapi/strapi'
+
+export default factories.createCoreController('api::content-offer.content-offer');
