+++ strapi/src/api/content-offer/routes/content-offer.ts
@@ -1,0 +1,7 @@
+/**
+ * content-offer router
+ */
+
+import { factories } from '@strapi/strapi';
+
+export default factories.createCoreRouter('api::content-offer.content-offer');
