+++ strapi/src/api/content-offer/services/content-offer.ts
@@ -1,0 +1,7 @@
+/**
+ * content-offer service
+ */
+
+import { factories } from '@strapi/strapi';
+
+export default factories.createCoreService('api::content-offer.content-offer');
