+++ strapi/src/api/event-page/content-types/event-page/schema.json
@@ -140,7 +140,6 @@
         "sections.product-feature-card",
         "relations.section-solutions-relation",
         "relations.sla-card-section",
-        "relations.white-paper-section",
         "sections.color-card-slider-section",
         "sections.icon-card-section-with-relation",
         "sections.info-details-section",
@@ -153,7 +152,8 @@
         "sections.video-with-text-section",
         "sections.feature-cards",
         "sections.content-carousel",
-        "sections.quote-section"
+        "sections.quote-section",
+        "sections.content-highlight-section"
       ]
     }
   }
