+++ strapi/src/api/homepage/content-types/homepage/schema.json
@@ -61,7 +61,6 @@
         "sections.product-feature-card",
         "relations.section-solutions-relation",
         "relations.sla-card-section",
-        "relations.white-paper-section",
         "sections.color-card-slider-section",
         "sections.icon-card-section-with-relation",
         "sections.info-details-section",
@@ -73,7 +72,8 @@
         "sections.video-with-text-section",
         "sections.quote-section",
         "sections.feature-cards",
-        "sections.content-carousel"
+        "sections.content-carousel",
+        "sections.content-highlight-section"
       ]
     }
   }
