+++ strapi/src/api/news-page/content-types/news-page/schema.json
@@ -79,7 +79,6 @@
       "components": [
         "sections.heading-with-link-container",
         "global.blog-block",
-        "sections.events-section-with-intro-and-cta",
         "sections.feature-cards",
         "sections.quote-section",
         "sections.content-carousel"
