+++ strapi/src/api/page-partner-and-product/content-types/page-partner-and-product/schema.json
@@ -71,7 +71,6 @@
         "sections.product-feature-card",
         "relations.section-solutions-relation",
         "relations.sla-card-section",
-        "relations.white-paper-section",
         "sections.color-card-slider-section",
         "sections.icon-card-section-with-relation",
         "sections.info-details-section",
@@ -84,7 +83,8 @@
         "sections.video-with-text-section",
         "sections.feature-cards",
         "sections.content-carousel",
-        "sections.quote-section"
+        "sections.quote-section",
+        "sections.content-highlight-section"
       ]
     }
   }
