--- strapi/src/api/white-paper/controllers/white-paper.ts
@@ -1,7 +1,0 @@
-/**
- * white-paper controller
- */
-
-import { factories } from '@strapi/strapi'
-
-export default factories.createCoreController('api::white-paper.white-paper');
