--- strapi/src/api/white-paper/routes/white-paper.ts
@@ -1,7 +1,0 @@
-/**
- * white-paper router
- */
-
-import { factories } from '@strapi/strapi';
-
-export default factories.createCoreRouter('api::white-paper.white-paper');
