--- strapi/src/api/white-paper/services/white-paper.ts
@@ -1,7 +1,0 @@
-/**
- * white-paper service
- */
-
-import { factories } from '@strapi/strapi';
-
-export default factories.createCoreService('api::white-paper.white-paper');
