--- strapi/src/components/relations/white-paper-section.json
@@ -1,20 +1,0 @@
-{
-  "collectionName": "components_relations_white_paper_sections",
-  "info": {
-    "displayName": "White Paper Section"
-  },
-  "options": {},
-  "attributes": {
-    "props": {
-      "type": "component",
-      "repeatable": false,
-      "component": "sections.section-props",
-      "required": true
-    },
-    "white_paper": {
-      "type": "relation",
-      "relation": "oneToOne",
-      "target": "api::white-paper.white-paper"
-    }
-  }
-}
