+++ strapi/src/middlewares/populate-sections-with-intro-and-localizations.ts
@@ -24,13 +24,6 @@ module.exports = (config, {strapi}) => {
                 cta: true,
               },
             },
-            'sections.projects-card-section-with-external-link': {
-              populate: {
-                projects: {
-                  populate: '*',
-                }
-              }
-            },
             'sections.product-feature-card': {
               populate: {
                 props: true,
@@ -78,10 +71,10 @@ module.exports = (config, {strapi}) => {
                 section_props: true,
               },
             },
-            'relations.white-paper-section': {
+            'sections.content-highlight-section': {
               populate: {
                 props: true,
-                white_paper: {
+                content_offer: {
                   populate: '*'
                 },
               },
